Calendar updates more frequently
